Jag är bäst

+++ src/Main/Math.java
@@ -6,8 +6,8 @@ public class Math {
         int sum = first + second;
         return sum;
     }
-    public int sub(int first, int second){
-        int sum = first - second;
+    public double sub(double first, double second){
+        double sum = first - second;
         return sum;
     }
 }
